UTs for AdaptiveApp
