+++ api/application_state_client/src/application_state_client.cpp
@@ -23,9 +23,4 @@ void ApplicationStateClient::ReportApplicationState(ApplicationState state)
   promise.wait(waitScope);
 }
 
-void ApplicationStateClientWrapper::ReportApplicationState(ApplicationStateManagement::ApplicationState state)
-{
-    m_client.ReportApplicationState(state);
-}
-
 } // namespace api
