+++ api/application_state_client/src/i_application_state_client_wrapper.cpp
@@ -1,0 +1,10 @@
+#include <i_application_state_client_wrapper.h>
+
+namespace api {
+
+void ApplicationStateClientWrapper::ReportApplicationState(ApplicationStateManagement::ApplicationState state)
+{
+    m_client.ReportApplicationState(state);
+}
+
+}
