+++ applications/AdaptiveApplication1/include/adaptive_app.hpp
@@ -3,6 +3,7 @@
 
 #include <i_adaptive_app.hpp>
 #include <i_state_factory.hpp>
+#include <i_application_state_client_wrapper.h>
 
 #include <vector>
